# openapi/openapi.yaml
...
openapi: 3.0.3
info:
    contact:
        name: MetaRGB
    description: REST API for MetaRGB microservices. Laravel-compatible JSON responses served via Kong → grpc-gateway → gRPC services.
    title: MetaRGB API
    version: 1.0.0
servers:
    - description: Kong API Gateway (local)
      url: http://localhost:8000
    - description: grpc-gateway direct (dev)
      url: http://localhost:8080
...
paths:
...
    /api/buy-requests:
        get:
            operationId: get_buy-requests
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create buy requests
            tags:
                - Features
        post:
            operationId: post_buy-requests
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create buy requests
            tags:
                - Features
    /api/buy-requests/{id}:
        delete:
            operationId: delete_buy-requests_id
            parameters:
                - description: Resource ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Manage buy request
            tags:
                - Features
        get:
            operationId: get_buy-requests_id
            parameters:
                - description: Resource ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Manage buy request
            tags:
                - Features
        patch:
            operationId: patch_buy-requests_id
            parameters:
                - description: Resource ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Manage buy request
            tags:
                - Features
        put:
            operationId: put_buy-requests_id
            parameters:
                - description: Resource ID
                  in: path
                  name: id
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: Manage buy request
            tags:
                - Features
...
    /api/features:
        get:
            operationId: get_features
            parameters:
                - description: Include building models in response
                  in: query
                  name: load_buildings
                  required: false
                  schema:
                    type: boolean
                - description: Bounding box corner points (min 4 values)
                  in: query
                  name: points
                  required: true
                  schema:
                    items:
                        type: string
                    type: array
                - description: Filter by user feature locations
                  in: query
                  name: user_features_location
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List features
            tags:
                - Features
...
    /api/features/buy/{feature}:
        post:
            operationId: post_features_buy_feature
            parameters:
                - description: Feature ID
                  in: path
                  name: feature
                  required: true
                  schema:
                    example: 1
                    format: int64
                    type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: BuyFeature
            tags:
                - Features
...
    /api/my-features:
        get:
            operationId: get_my-features
            parameters:
                - description: Page number for pagination
                  in: query
                  name: page
                  required: false
                  schema:
                    type: integer
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List authenticated user features
            tags:
                - Features
...
    /api/sell-requests:
        get:
            operationId: get_sell-requests
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create sell requests
            tags:
                - Features
        post:
            operationId: post_sell-requests
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - BearerAuth: []
            summary: List or create sell requests
            tags:
                - Features
...
    /api/store:
        get:
            operationId: get_store
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: List store packages
            tags:
                - Financial
        post:
            operationId: post_store
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                    multipart/form-data:
                        schema:
                            type: object
            responses:
                "200":
                    description: Successful response
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description: Unauthenticated
                "422":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ValidationError'
                    description: Validation error
            security:
                - {}
                - BearerAuth: []
            summary: GetStorePackages
            tags:
                - Store
...
components:
    schemas:
        Error:
            properties:
                message:
                    type: string
            type: object
        ValidationError:
            properties:
                errors:
                    additionalProperties:
                        items:
                            type: string
                        type: array
                    type: object
                message:
                    type: string
            type: object
    securitySchemes:
        BearerAuth:
            bearerFormat: Token
            description: 'Laravel Sanctum-style bearer token. Header: Authorization: Bearer {token}'
            scheme: bearer
            type: http